pyproject.toml
[project]
name = "weather-forecast-at-the-cloud"
version = "0.1.0"
description = "Machine learning applied to weather forecasting, with data from REM San Luis"
readme = "README.md"
authors = [
    { name = "DanielBerns", email = "daniel.w.berns@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.5",
    "numpy>=2.1.3",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
    "tensorflow>=2.19.0",
]

[project.scripts]
weather-forecast-at-the-cloud = "weather_forecast_at_the_cloud:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
